requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.21.0
qdrant-client>=1.6.0
httpx>=0.25.0
pillow>=10.0.0
python-multipart>=0.0.6
fashion-clip>=0.3.0
